restructure registration/login
