+++ src/App.tsx
@@ -4,9 +4,9 @@ import './assets/fonts/fonts.css';
 import { Layout } from './components/UI/Layout/Layout';
 import { Tag } from './containers/Tag/Tag';
 import { TagPage } from './components/pages/TagPage/TagPage';
-import { Registration } from './components/pages/Registration/Registration';
-import { ConfirmOTP } from './components/pages/ConfirmOTP/ConfirmOTP';
-import { Login } from './components/pages/Login/Login';
+import { Registration } from './containers/Registration/Registration';
+import { ConfirmOTP } from './containers/ConfirmOTP/ConfirmOTP';
+import { Login } from './containers/Login/Login';
 import MessageTemplatePage from './components/pages/MessageTemplatePage/MessageTemplatePage';
 import { MessageTemplate } from './containers/MessageTemplate/MessageTemplate';
 import Chat from './containers/Chat/Chat';
